Split ThreatModel into its own library
